backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=2.1.0
scipy>=1.13.0
yfinance>=0.2.40
python-multipart>=0.0.9
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pydantic>=2.9.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
